requirements.txt
# PDF Processing
PyMuPDF==1.23.8  # For PDF text extraction and manipulation
pdf2image==1.16.3  # For converting PDFs to images when needed

# OCR and Text Processing
google-cloud-vision==3.4.4  # For Gemini 2.0 OCR capabilities
python-Levenshtein==0.23.0  # For fuzzy text matching
nltk==3.8.1  # For text processing and analysis

# Database and ORM
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9  # PostgreSQL adapter
alembic==1.13.1  # Database migrations
pgvector==0.2.4  # Vector operations support for PostgreSQL

# API and Web Framework
fastapi==0.109.0
uvicorn==0.25.0  # ASGI server
python-multipart==0.0.6  # For handling file uploads
pydantic==2.5.3  # Data validation
pydantic-settings==2.1.0  # Settings management

# Graph Processing
networkx==3.2.1  # For theorem dependency graphs
matplotlib==3.8.2  # For graph visualization

# Machine Learning and Embeddings
torch==2.1.2  # For neural network operations
transformers==4.36.2  # For text embeddings
sentence-transformers==2.2.2  # For semantic similarity

# Testing and Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1  # Code formatting
isort==5.13.2  # Import sorting
mypy==1.8.0  # Type checking

# Container and Infrastructure
docker==7.0.0
kubernetes==28.1.0  # For K8s orchestration